Attempt read from indicated starting offset from font header when table overlap occurs.
Tested against CenturyOldStyle-Light Bold (alltype).

Whether or not this should be done in general...
But it seems to work okay. It's a case of 'what to believe' - is the header wrong and the current offset correct, or is the indicated offset correct?

In this case, we're preferring the offset indicated in the header over the current position.